# *: use 8 byte MysqlTime format