--- a/distsql/distsql.go
+++ b/distsql/distsql.go
@@ -166,9 +166,7 @@ func canUseChunkRPC(ctx sessionctx.Context) bool {
 	return true
 }
 
-var supportedAlignment = !(unsafe.Sizeof(types.MysqlTime{}) != 16 ||
-	unsafe.Sizeof(types.Time{}) != 20 ||
-	unsafe.Sizeof(types.MyDecimal{}) != 40)
+var supportedAlignment = unsafe.Sizeof(types.MyDecimal{}) == 40
 
 // checkAlignment checks the alignment in current system environment.
 // The alignment is influenced by system, machine and Golang version.