PCP #14191 
PCP #14192 

Signed-off-by: TennyZhuang <zty0826@gmail.com>

<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

#14191 
#14192 

### What is changed and how it works?

- [x] Reduce MysqlTime to 8 byte.
- [x] Reduce Time to 8 byte.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Code changes

 - Has exported function/method change
 - Has exported variable/fields change

Side effects

Related changes

Release note

 - Reduce the Time structure size

Benchmark

run

```bash
make vectorized-bench VB_FILE=Time VB_FUNC=builtinAddDateDatetimeStringSig
```

on master branch:

```
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc-8         	    7353	    154228 ns/op	   16656 B/op	     201 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc-8      	    4999	    212088 ns/op	   18224 B/op	     299 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#01-8      	    6085	    178985 ns/op	   21640 B/op	     261 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#01-8   	    4561	    270220 ns/op	   22480 B/op	     366 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#02-8      	    8046	    151590 ns/op	   17480 B/op	     211 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#02-8   	    5074	    209606 ns/op	   18344 B/op	     319 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#03-8      	    4267	    275043 ns/op	   43373 B/op	     521 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#03-8   	    2664	    390195 ns/op	   43800 B/op	     626 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#04-8      	    4551	    299354 ns/op	   41710 B/op	     501 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#04-8   	    3068	    364122 ns/op	   42069 B/op	     601 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#05-8      	    4888	    250683 ns/op	   43684 B/op	     526 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#05-8   	    3188	    345586 ns/op	   44100 B/op	     630 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#06-8      	    4636	    240326 ns/op	   42021 B/op	     506 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#06-8   	    3114	    402775 ns/op	   42554 B/op	     606 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#07-8      	    4980	    232038 ns/op	   40776 B/op	     491 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#07-8   	    3277	    355382 ns/op	   41552 B/op	     588 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#08-8      	    4104	    275428 ns/op	   44100 B/op	     531 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#08-8   	    2798	    375312 ns/op	   44520 B/op	     636 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#09-8      	    4369	    245849 ns/op	   38736 B/op	     499 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#09-8   	    3354	    307135 ns/op	   42003 B/op	     601 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#10-8      	    4413	    253583 ns/op	   39805 B/op	     508 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#10-8   	    3244	    327015 ns/op	   43126 B/op	     612 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#11-8      	    4376	    258448 ns/op	   40520 B/op	     712 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#11-8   	    3469	    324351 ns/op	   42140 B/op	     814 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#12-8      	    4857	    268345 ns/op	   39244 B/op	     499 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#12-8   	    3180	    315699 ns/op	   40839 B/op	     600 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#13-8      	    4335	    366588 ns/op	   37807 B/op	     658 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#13-8   	    3374	    462400 ns/op	   39284 B/op	     753 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#14-8      	    3853	    260113 ns/op	   44414 B/op	     759 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#14-8   	    3151	    382425 ns/op	   46058 B/op	     864 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#15-8      	    4288	    242266 ns/op	   37409 B/op	     471 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#15-8   	    4117	    346343 ns/op	   38923 B/op	     567 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#16-8      	    4273	    284653 ns/op	   38697 B/op	     691 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#16-8   	    3832	    328215 ns/op	   39419 B/op	     794 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#17-8      	    4842	    252504 ns/op	   40963 B/op	     705 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#17-8   	    4123	    307438 ns/op	   41716 B/op	     801 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#18-8      	    3874	    344169 ns/op	   65083 B/op	    1010 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#18-8   	    2564	    454909 ns/op	   66003 B/op	    1117 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#19-8      	    3686	    318695 ns/op	   60249 B/op	     926 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#19-8   	    2565	    435963 ns/op	   61112 B/op	    1023 allocs/op
BenchmarkVectorizedBuiltinTimeFunc-8                                                                 	1000000000	         0.0268 ns/op	       0 B/op	       0 allocs/op
```

after optimize:

```
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc-8         	    7204	    157268 ns/op	   16656 B/op	     201 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc-8      	    5157	    211539 ns/op	   18320 B/op	     305 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#01-8      	    7416	    155400 ns/op	   16232 B/op	     196 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#01-8   	    5152	    208961 ns/op	   17096 B/op	     304 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#02-8      	    7531	    153779 ns/op	   19560 B/op	     236 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#02-8   	    5029	    215634 ns/op	   20432 B/op	     345 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#03-8      	    3915	    263461 ns/op	   44999 B/op	     541 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#03-8   	    2556	    398462 ns/op	   45493 B/op	     652 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#04-8      	    4620	    243280 ns/op	   42509 B/op	     511 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#04-8   	    3032	    365399 ns/op	   42894 B/op	     612 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#05-8      	    5316	    209939 ns/op	   42020 B/op	     506 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#05-8   	    3444	    331576 ns/op	   42420 B/op	     606 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#06-8      	    5604	    212238 ns/op	   38693 B/op	     466 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#06-8   	    3519	    321116 ns/op	   39184 B/op	     558 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#07-8      	    4890	    238576 ns/op	   44104 B/op	     531 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#07-8   	    2898	    364694 ns/op	   44944 B/op	     636 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#08-8      	    4347	    250215 ns/op	   45764 B/op	     551 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#08-8   	    3052	    377183 ns/op	   46200 B/op	     660 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#09-8      	    4686	    230060 ns/op	   37752 B/op	     482 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#09-8   	    3858	    295115 ns/op	   40843 B/op	     579 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#10-8      	    4393	    248051 ns/op	   40783 B/op	     522 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#10-8   	    3372	    318209 ns/op	   44058 B/op	     624 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#11-8      	    4752	    239136 ns/op	   38446 B/op	     676 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#11-8   	    3642	    298288 ns/op	   39966 B/op	     774 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#12-8      	    4323	    254808 ns/op	   41856 B/op	     540 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#12-8   	    3358	    321328 ns/op	   43601 B/op	     650 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#13-8      	    4950	    235124 ns/op	   37430 B/op	     660 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#13-8   	    3562	    291766 ns/op	   39033 B/op	     760 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#14-8      	    4626	    227698 ns/op	   40457 B/op	     694 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#14-8   	    3663	    292774 ns/op	   42010 B/op	     792 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#15-8      	    5064	    216693 ns/op	   35129 B/op	     447 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#15-8   	    4140	    272675 ns/op	   36700 B/op	     544 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#16-8      	    4293	    255224 ns/op	   42103 B/op	     727 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#16-8   	    3396	    320953 ns/op	   42888 B/op	     829 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#17-8      	    5140	    223471 ns/op	   39876 B/op	     691 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#17-8   	    3855	    286012 ns/op	   40645 B/op	     790 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#18-8      	    3902	    281779 ns/op	   59049 B/op	     915 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#18-8   	    2824	    393541 ns/op	   59791 B/op	    1012 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-VecBuiltinFunc#19-8      	    3594	    303328 ns/op	   60913 B/op	     944 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateDatetimeStringSig-NonVecBuiltinFunc#19-8   	    2611	    422433 ns/op	   61742 B/op	    1042 allocs/op
BenchmarkVectorizedBuiltinTimeFunc-8                                                                 	1000000000	         0.0247 ns/op	       0 B/op	       0 allocs/op
```

Diff:

https://www.diffchecker.com/qDShBzIx