@TennyZhuang 2020-01-06
I have updated a benchmark result in PR description.