@wshwsh12 2020-01-07
/run-all-tests tikv=pr/6418