@breeswish 2020-01-12
@TennyZhuang Looks like you need to fix conflict (merge master) in order to proceed the test due to some PRIVILEGE related changes in master.