@TennyZhuang 2020-01-12
@sre-bot  /run-all-tests tikv=pr/6418 tidb-test=pr/975