@TennyZhuang 2020-01-15
@wshwsh12 the bench will failed because we   need bench between tidb/14278+tikv/6148 and tidb/master+tikv/master, currently the bench command cannot support it. @breeswish has some test result by manually.