@wshwsh12 2020-01-15
/run-all-tests tikv=pr/6418 tidb-test=pr/975