@TennyZhuang 2020-01-21
@sre-bot /bench +tpch tikv=pr/6418