@TennyZhuang 2020-01-22
@sre-bot /bench +tpch tikv=pr/6418