@TennyZhuang 2020-01-22
@sre-bot  /run-all-tests tikv=pr/6418 tidb-test=pr/975