@lonng 2020-01-31
/run-all-tests tikv=pr/6418 tidb-test=pr/975