@lonng 2020-01-31
/run-common-test tikv=pr/6418 tidb-test=pr/975